[Change] Move initWeb3Onboard to constants
